♻️  [REFACTOR] 회원가입 및 프로필 설정 페이지 추가
- SignupPage 및 ProfileSetupPage 컴포넌트 추가
- AuthHeader 및 IconInput 컴포넌트 생성하여 재사용성 향상
- 프로필 설정에 필요한 상수 파일 추가 및 관련 컴포넌트 구성
- 기존 SignupPage 및 ProfileSetupPage 삭제 후 새로운 구조로 통합
