+++ src/App.tsx
@@ -10,8 +10,8 @@ import NavigationHandler from './components/NavigationHandler';
 const Index = lazy(() => import("./pages/Index"));
 const LandingPage = lazy(() => import("./pages/landing/LandingPage"));
 const LoginPage = lazy(() => import("./pages/login/LoginPage"));
-const SignupPage = lazy(() => import("./pages/SignupPage"));
-const ProfileSetupPage = lazy(() => import("./pages/ProfileSetupPage"));
+const SignupPage = lazy(() => import("./pages/signup/SignupPage"));
+const ProfileSetupPage = lazy(() => import("./pages/setup-profile/ProfileSetupPage"));
 const MatchingPage = lazy(() => import("./pages/matching/MatchingPage"));
 const WaitingRoomPage = lazy(() => import("./pages/WaitingRoomPage"));
 const ScreenShareSetupPage = lazy(() => import("./pages/ScreenShareSetupPage"));
