+++ src/pages/setup-profile/constants.ts
@@ -1,0 +1,49 @@
+export const programmingLanguages = [
+    "Python",
+    "Java",
+    "JavaScript",
+    "C++",
+    "C",
+    "C#",
+    "Go",
+    "Rust",
+    "Kotlin",
+    "Swift",
+    "TypeScript",
+    "Ruby",
+  ];
+  
+  export const solvedacTiers = [
+    "Unrated",
+    "Bronze V",
+    "Bronze IV",
+    "Bronze III",
+    "Bronze II",
+    "Bronze I",
+    "Silver V",
+    "Silver IV",
+    "Silver III",
+    "Silver II",
+    "Silver I",
+    "Gold V",
+    "Gold IV",
+    "Gold III",
+    "Gold II",
+    "Gold I",
+    "Platinum V",
+    "Platinum IV",
+    "Platinum III",
+    "Platinum II",
+    "Platinum I",
+    "Diamond V",
+    "Diamond IV",
+    "Diamond III",
+    "Diamond II",
+    "Diamond I",
+    "Ruby V",
+    "Ruby IV",
+    "Ruby III",
+    "Ruby II",
+    "Ruby I",
+    "Master",
+  ];
